Introduce Order#callback_data not to be confused with Order#data
callback_data holds either a string or a Hash that is returned to the
callback_url in the http param, whereas #data is simply something we only
store in the DB. This distinction is important because sometimes
we want to automatically fill the #data attr with relevant info, like
exchange rate at the time of purchase but a merchant doesn't expect us
to return it back to him in the callback.

More importantly, when we had only #data attrribute, it conflicted with
what the merchant sent to us. So, for example, we might have overwritten it
with exchange rate data even though merchant sends us a simple string he
expects to be returned in the callback.